Manage GitHub issues comments from this repo
Instead of commenting on the GitHub issue from outside entity that
initiates the build, we post comments from this repository.

The advantage is that, when we receive the build request, we
can receive the APT package name and the GitHub issue number.

The basic workflow is like this:

* Run our check on a docker container on the VM, then depending
on the result, we exit the ssh session.
* Based on the exit status, we create a PR or post comments