Add run_tests.rb [skip ci]
This script consults env var REPO to look for issues with title
matching the APT whitelist requests.

This script must be run in the root directory of this git repo.

If the issue has the label 'apt-whitelist-check-run', it will be
skipped.

If the title contains "source", then an appropriate label is
added and moves on.

If the title is ambiguous, a comment to request clarification is
posted, and an appropriate label is added.

Finally, when the title is something that we can act on, we make a
new commit to *this* repository, pushes it to github, and that will
trigger the necessary check.

At the end of the check, the label 'apt-whitelist-check-run' is
added to the issue.